requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
gunicorn
python-jose[cryptography]==1.3.0
python-multipart
SQLAlchemy
python-dotenv
httpx
email-validator
cryptography
psutil
GitPython